chore: allocated types, constants and utility functions to separate files
