+++ app/leaderboard/_components/LeaderboardWrapper.tsx
@@ -1,7 +1,7 @@
 import { getLeaderboardData } from "../../api/leaderboard/functions";
 import { parseDateRangeSearchParam } from "@/lib/utils";
 import Leaderboard from "./Leaderboard";
-import { LeaderboardPageProps } from "../page";
+import { LeaderboardPageProps } from "@/lib/types";
 
 export default async function LeaderboardWrapper({
   searchParams,
