+++ components/contributors/TopContributor.tsx
@@ -1,17 +1,8 @@
-import { LeaderboardAPIResponse } from "@/app/api/leaderboard/functions";
+import { LeaderboardAPIResponse } from "@/lib/types";
 import Image from "next/image";
 import Link from "next/link";
-
-export const TOP_CONTRIBUTOR_CATEGORIES = {
-  eod_update: "EOD Updates",
-  pr_opened: "Pull Requests Opened",
-  pr_merged: "Pull Requests Merged",
-  pr_reviewed: "Pull Requests Reviewed",
-  issue_opened: "Issues Opened",
-  comment_created: "Comments Created",
-};
-
-export type TopContributorCategoryKey = keyof typeof TOP_CONTRIBUTOR_CATEGORIES;
+import { TOP_CONTRIBUTOR_CATEGORIES } from "@/lib/const";
+import { TopContributorCategoryKey } from "@/lib/types";
 
 export default function InfoCard({
   data,
